pyproject.toml
[project]
name = "42bingo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0.3",
    "reportlab>=4.4.7",
    "ruff>=0.14.10",
]
